Made things look pretty
I take so much time to get colors and the nav bar right…

I need to practise just making nav bars right and look great and fit on
the screen properly. I’m learning.. but slowly.

Also chose 2 fonts that I enjoy for the site. Sans serif for the win.